Updated the Dockerfile and moved the database files to admin-module
